Estructuracion de proyecto
